extract the seed function from casper to make it easier to maintain
